Refactor help components and update usage
- Extracted financial help guidance into a new HowDoIGetHelp component and simplified NeedHelp only include NeedHelp info.
- Updated DebtDetails, DebtLettersSummary, and ResolveDebtPage to use the new components, improving separation of concerns and maintainability. Also made markup and prop improvements in HowDoIDispute and HowDoIPay.
- Moved modal button to bottom of the page on the debt detail page with OTTP turned off
