+++ src/applications/combined-debt-portal/debt-letters/containers/DebtDetails.jsx
@@ -3,10 +3,8 @@ import { Link, useLocation } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import last from 'lodash/last';
 import { VaBreadcrumbs } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
-import { CONTACTS } from '@department-of-veterans-affairs/component-library/contacts';
 import { head } from 'lodash';
 import HowDoIPay from '../components/HowDoIPay';
-import NeedHelp from '../components/NeedHelp';
 import HistoryTable from '../components/HistoryTable';
 import {
   setPageFocus,
@@ -24,6 +22,8 @@ import DebtDetailsCard from '../components/DebtDetailsCard';
 import PaymentHistoryTable from '../components/PaymentHistoryTable';
 import useHeaderPageTitle from '../../combined/hooks/useHeaderPageTitle';
 import Modals from '../../combined/components/Modals';
+import HowDoIGetHelp from '../components/HowDoIGetHelp';
+import NeedHelp from '../components/NeedHelp';
 
 const DebtDetails = () => {
   const { selectedDebt, debts } = useSelector(
@@ -232,34 +232,20 @@ const DebtDetails = () => {
             ) : null}
           </>
         )}
-        <Modals title="Notice of rights and responsibilities" id="notice-modal">
-          <Modals.Rights />
-        </Modals>
 
-        {oneThingPerPageActive ? (
-          <va-need-help id="needHelp">
-            <div slot="content">
-              <p>
-                If you have any questions about your benefit overpayment.
-                Contact us online through{' '}
-                <a href="https://ask.va.gov/">Ask VA</a> or call the Debt
-                Management Center at <va-telephone contact={CONTACTS.DMC} /> (
-                <va-telephone contact="711" tty="true" />
-                ). For international callers, use{' '}
-                <va-telephone contact={CONTACTS.DMC_OVERSEAS} international />.
-                We’re here Monday through Friday, 7:30 a.m. to 7:00 p.m. ET.
-              </p>
-            </div>
-          </va-need-help>
-        ) : (
+        {!oneThingPerPageActive && (
           <>
             <HowDoIPay userData={howToUserData} />
-            <NeedHelp
+            <HowDoIGetHelp
               showVHAPaymentHistory={false}
               showOneThingPerPage={oneThingPerPageActive}
             />
           </>
         )}
+        <Modals title="Notice of rights and responsibilities" id="notice-modal">
+          <Modals.Rights />
+        </Modals>
+        <NeedHelp />
       </div>
     </article>
   );
