+++ src/applications/combined-debt-portal/debt-letters/containers/ResolveDebtPage.jsx
@@ -10,6 +10,7 @@ import { setPageFocus } from '../../combined/utils/helpers';
 import useHeaderPageTitle from '../../combined/hooks/useHeaderPageTitle';
 import { getCurrentDebt } from '../utils/page';
 import { deductionCodes } from '../const/deduction-codes';
+import HowDoIGetHelp from '../components/HowDoIGetHelp';
 
 const ResolveDebtPage = ({ match }) => {
   const { selectedDebt, debts } = useSelector(
@@ -41,7 +42,7 @@ const ResolveDebtPage = ({ match }) => {
   }, []);
 
   return (
-    <>
+    <article>
       <VaBreadcrumbs
         breadcrumbList={[
           { href: '/', label: 'VA.gov Home' },
@@ -72,9 +73,10 @@ const ResolveDebtPage = ({ match }) => {
         </p>
         <va-on-this-page class="medium-screen:vads-u-margin-top--0" />
         <HowDoIPay userData={howToUserData} />
-        <NeedHelp showVHAPaymentHistory={showVHAPaymentHistory} />
+        <HowDoIGetHelp showVHAPaymentHistory={showVHAPaymentHistory} />
+        <NeedHelp />
       </div>
-    </>
+    </article>
   );
 };
